Basic Functions

some basic functions in numerical analysis, you could use.

Usage

Functions

NumericalAnalysis.NthDerivative — Method

NthDrivative(f::Function, x::Real, n::Int)

return the nth derivative value at x

NumericalAnalysis.TaylorPolynomials — Method

TaylorPolynomials(f::Function, x::Real, x₀::Real, n::Int)

return the value of nth Taylor Ploynomial at x.

NumericalAnalysis.absolute_error — Method

absolute_error(x, x̂)

return the Absolute error of x and x̂

NumericalAnalysis.relative_error — Method

relative_error(x, x̂)

The relative error between X1 and X2

NumericalAnalysis.ξ — Method

ξ(x₀, x)

return the number ξ(x) bwteen x₀ and x.