Solutions of Equations in One variable

The most basic problems of Numerical approximation the root-finding problem. $ f(x) = 0 $

Introduction

The Bisection Method

The Bisection technique based on the Intermediate Value Theorem, called the Binary-search method

First, suppose f is continuous function defined on the interval [a, b], with $f(a)\cdot f(b) <0$,The Intermediate Value Theorem tell us that $\exist p \in (a, b)$ with $f(p) = 0$.

To begin:

Set $a_1 = a$ and $b_1=b$, and $p_1 = \frac{a_1+b_1}{2}$

if $f(p_1) = 0$, So done.

If $f(p_1) \ne 0$, then $f(p_1)$ has the same sign as either $f(a)$ and $f(b)$

If $f(p_1)$ and $f(a)$ have the same sign, root in $(p_1, b_1)$, change: $a_2 = p_1$ and $b_2 = b_1$.

if $f(p_1)$ and $f(b)$ have the same sign, root in $(a_1, p_1)$, change: $a_2 = a_1$ and $b_2=p_1$.

  using NumericalAnalysis: SEq1
  SEq1.Bisection(sin, π/2, 3π/2)

3.141688527389036

and the real answers

π

π = 3.1415926535897...

Method

NumericalAnalysis.SEq1 — Module

SEq1

Solutions of Equations in One Variable

NumericalAnalysis.SEq1.Bisection — Function

Bisection(f::Function, a::Real, b::Real, TOL::Float64=10^(-4), N::Int=10)

Use Bisection To find a solution to. $f(x) = 0$ You should use the function, and a interval [a, b], tolerance TOL, maximum number of iterations N₀. Then you will get approximate solution $p$ or can't find in maximum N and return -1.